WE-578 remove flicker on opening the website
